+++ academy-ui/src/components/bytes/View/SwiperByteView/SwiperByteStepperItemContent.tsx
@@ -1,6 +1,6 @@
 import UserDiscord from '@/components/app/Form/UserDiscord';
 import { QuestionSection } from '@/components/bytes/View/QuestionSection';
-import { UseGenericViewByteHelper } from '@/components/bytes/View/useGenericViewByte';
+import { UseViewByteHelper } from '@/components/bytes/View/useViewByteInModal';
 import {
   ByteQuestionFragmentFragment,
   ByteUserDiscordConnectFragmentFragment,
@@ -22,7 +22,7 @@ interface ByteStepperItemContentProps {
   byte: ByteDto;
   step: ByteStepDto;
   space: SpaceWithIntegrationsFragment;
-  viewByteHelper: UseGenericViewByteHelper;
+  viewByteHelper: UseViewByteHelper;
   setByteSubmitted: (submitted: boolean) => void;
   renderer: marked.Renderer;
   activeStepOrder: number;
