+++ academy-ui/src/components/bytes/View/useViewByteInModal.ts
@@ -1,5 +1,6 @@
-import { checkIfQuestionIsComplete, checkIfUserInputIsComplete, updateItemValue } from '@/components/bytes/View/viewByteHelper';
 import { ByteDto, ByteStepDto } from '@/types/bytes/ByteDto';
+import { ByteStepItem, Question, UserInput } from '@/types/stepItems/stepItemDto';
+import { isQuestion, isUserInput } from '@dodao/web-core/types/deprecated/helpers/stepItemTypes';
 import { useNotificationContext } from '@dodao/web-core/ui/contexts/NotificationContext';
 import { ByteSubmissionInput, SpaceWithIntegrationsFragment } from '@/graphql/generated/generated-types';
 import { Session } from '@dodao/web-core/types/auth/Session';
@@ -7,6 +8,7 @@ import { LocalStorageKeys } from '@dodao/web-core/types/deprecated/models/enums'
 import { ByteSubmissionError } from '@dodao/web-core/types/errors/error';
 import { StepItemResponse, StepResponse, TempByteSubmission } from '@/utils/byte/TempByteSubmission';
 import getBaseUrl from '@dodao/web-core/utils/api/getBaseURL';
+import isEqual from 'lodash/isEqual';
 import union from 'lodash/union';
 import { useSession } from 'next-auth/react';
 import { useState } from 'react';
@@ -20,6 +22,31 @@ export interface UseViewByteInModalArgs {
   stepOrder: number;
   fetchByteFn: (byteId: string) => Promise<ByteDto>;
 }
+
+export interface UseViewByteHelper {
+  initialize: () => Promise<void>;
+  activeStepOrder: number;
+  canNavigateToNext: (step: ByteStepDto) => boolean;
+  isPristine: (stepUuid: string) => boolean;
+  errors: ByteSubmissionError;
+  getStepSubmission: (stepUuid: string) => StepResponse | undefined;
+  getStepItemSubmission: (stepUuid: string, stepItemUuid: string) => StepItemResponse | undefined;
+  goToNextStep: (currentStep: ByteStepDto) => void;
+  goToPreviousStep: (currentStep: ByteStepDto) => void;
+  byteLoaded: boolean;
+  byteRef: ByteDto;
+  byteSubmission: TempByteSubmission;
+  byteSubmitting: boolean;
+  isUserInputComplete: (stepUuid: string) => boolean;
+  isQuestionAnswered: (stepUuid: string) => boolean;
+  isValidToSubmit: () => boolean;
+  selectAnswer: (stepUuid: string, questionUuid: string, selectedAnswers: string[]) => void;
+  setActiveStep: (order: number) => void;
+  submitByte: () => Promise<boolean>;
+  setUserInput: (stepUuid: string, userInputUuid: string, userInput: string) => void;
+  setUserDiscord: (stepUuid: string, userDiscordUuid: string, userDiscordId: string) => void;
+}
+
 export function useViewByteInModal({ space, byteId, stepOrder, fetchByteFn }: UseViewByteInModalArgs): UseViewByteHelper {
   const { data: sessionData } = useSession();
   const session: Session | null = sessionData as Session | null;
@@ -30,6 +57,8 @@ export function useViewByteInModal({ space, byteId, stepOrder, fetchByteFn }: Us
   const [byteStepsMap, setByteStepsMap] = useState<{ [uuid: string]: ByteStepDto }>({});
   const [byteSubmitting, setByteSubmitting] = useState<boolean>(false);
   const [errors, setErrors] = useState<ByteSubmissionError>({});
+
+  const [pristineSteps, setPristineSteps] = useState<{ [stepUuid: string]: boolean }>({});
 
   const [byteSubmission, setByteSubmission] = useState<TempByteSubmission>({
     isPristine: true,
@@ -85,6 +114,19 @@ export function useViewByteInModal({ space, byteId, stepOrder, fetchByteFn }: Us
   function getStepItemSubmission(stepUuid: string, stepItemUuid: string): StepItemResponse | undefined {
     const stepSubmission = getStepSubmission(stepUuid);
     return stepSubmission?.itemResponsesMap[stepItemUuid];
+  }
+
+  function canNavigateToNext(step: ByteStepDto): boolean {
+    const answeredCorrectly = step.stepItems.filter(isQuestion).every((stepItem: ByteStepItem) => {
+      const question = stepItem as Question;
+      return isEqual(question.answerKeys.sort(), ((getStepItemSubmission(step.uuid, stepItem.uuid) as string[]) || []).sort());
+    });
+
+    return isQuestionAnswered(step.uuid) && isUserInputComplete(step.uuid) && answeredCorrectly;
+  }
+
+  function isPristine(stepUuid: string) {
+    return pristineSteps[stepUuid];
   }
 
   function goToNextStep(currentStep: ByteStepDto) {
@@ -216,6 +258,8 @@ export function useViewByteInModal({ space, byteId, stepOrder, fetchByteFn }: Us
   return {
     initialize,
     activeStepOrder,
+    canNavigateToNext,
+    isPristine,
     errors,
     getStepSubmission,
     getStepItemSubmission,
@@ -236,24 +280,68 @@ export function useViewByteInModal({ space, byteId, stepOrder, fetchByteFn }: Us
   };
 }
 
-export interface UseViewByteHelper {
-  initialize: () => Promise<void>;
-  activeStepOrder: number;
-  errors: ByteSubmissionError;
-  getStepSubmission: (stepUuid: string) => StepResponse | undefined;
-  getStepItemSubmission: (stepUuid: string, stepItemUuid: string) => StepItemResponse | undefined;
-  goToNextStep: (currentStep: ByteStepDto) => void;
-  goToPreviousStep: (currentStep: ByteStepDto) => void;
-  byteLoaded: boolean;
-  byteRef: ByteDto;
-  byteSubmission: TempByteSubmission;
-  byteSubmitting: boolean;
-  isUserInputComplete: (stepUuid: string) => boolean;
-  isQuestionAnswered: (stepUuid: string) => boolean;
-  isValidToSubmit: () => boolean;
-  selectAnswer: (stepUuid: string, questionUuid: string, selectedAnswers: string[]) => void;
-  setActiveStep: (order: number) => void;
-  submitByte: () => Promise<boolean>;
-  setUserInput: (stepUuid: string, userInputUuid: string, userInput: string) => void;
-  setUserDiscord: (stepUuid: string, userDiscordUuid: string, userDiscordId: string) => void;
+export function checkIfUserInputIsComplete(
+  stepUuid: string,
+  byteStepsMap: {
+    [p: string]: ByteStepDto;
+  },
+  getStepSubmission: (stepUuid: string) => StepResponse | undefined
+) {
+  if (stepUuid === LAST_STEP_UUID) return true;
+  const step = byteStepsMap[stepUuid];
+  const stepSubmission = getStepSubmission(stepUuid);
+  if (!step) {
+    console.error(`no step with uuid - ${stepUuid} found in`, byteStepsMap);
+  }
+
+  const respondedToAllInputs = step.stepItems
+    .filter(isUserInput)
+    .filter((item) => (item as UserInput).required)
+    .every((userInput) => (stepSubmission?.itemResponsesMap?.[userInput.uuid] as string)?.length);
+
+  return respondedToAllInputs;
+}
+
+export function checkIfQuestionIsComplete(
+  stepUuid: string,
+  byteStepsMap: {
+    [p: string]: ByteStepDto;
+  },
+  getStepSubmission: (stepUuid: string) => StepResponse | undefined
+) {
+  if (stepUuid === LAST_STEP_UUID) return true;
+  const step = byteStepsMap[stepUuid];
+  if (!step) {
+    console.error(`no step with uuid - ${stepUuid} found in`, byteStepsMap);
+  }
+  const stepSubmission = getStepSubmission(stepUuid);
+
+  const allQuestionsAnswered = step.stepItems
+    .filter(isQuestion)
+    .every((question) => (stepSubmission?.itemResponsesMap?.[question.uuid] as string[] | undefined)?.length);
+
+  return allQuestionsAnswered;
+}
+
+export function updateItemValue(
+  stepUuid: string,
+  itemUuid: string,
+  itemValue: string | string[],
+  setByteSubmission: (fn: (submission: TempByteSubmission) => TempByteSubmission) => void
+) {
+  setByteSubmission((prevByteSubmission: TempByteSubmission) => {
+    return {
+      ...prevByteSubmission,
+      stepResponsesMap: {
+        ...prevByteSubmission.stepResponsesMap,
+        [stepUuid]: {
+          ...prevByteSubmission.stepResponsesMap?.[stepUuid],
+          itemResponsesMap: {
+            ...prevByteSubmission.stepResponsesMap?.[stepUuid]?.itemResponsesMap,
+            [itemUuid]: itemValue,
+          },
+        },
+      },
+    };
+  });
 }
