--- academy-ui/src/components/bytes/View/viewByteHelper.ts
@@ -1,71 +1,0 @@
-import { LAST_STEP_UUID } from '@/components/bytes/View/useGenericViewByte';
-import { ByteStepDto } from '@/types/bytes/ByteDto';
-import { UserInput } from '@/types/stepItems/stepItemDto';
-import { isQuestion, isUserInput } from '@dodao/web-core/types/deprecated/helpers/stepItemTypes';
-import { StepResponse, TempByteSubmission } from '@/utils/byte/TempByteSubmission';
-
-export function checkIfUserInputIsComplete(
-  stepUuid: string,
-  byteStepsMap: {
-    [p: string]: ByteStepDto;
-  },
-  getStepSubmission: (stepUuid: string) => StepResponse | undefined
-) {
-  if (stepUuid === LAST_STEP_UUID) return true;
-  const step = byteStepsMap[stepUuid];
-  const stepSubmission = getStepSubmission(stepUuid);
-  if (!step) {
-    console.error(`no step with uuid - ${stepUuid} found in`, byteStepsMap);
-  }
-
-  const respondedToAllInputs = step.stepItems
-    .filter(isUserInput)
-    .filter((item) => (item as UserInput).required)
-    .every((userInput) => (stepSubmission?.itemResponsesMap?.[userInput.uuid] as string)?.length);
-
-  return respondedToAllInputs;
-}
-
-export function checkIfQuestionIsComplete(
-  stepUuid: string,
-  byteStepsMap: {
-    [p: string]: ByteStepDto;
-  },
-  getStepSubmission: (stepUuid: string) => StepResponse | undefined
-) {
-  if (stepUuid === LAST_STEP_UUID) return true;
-  const step = byteStepsMap[stepUuid];
-  if (!step) {
-    console.error(`no step with uuid - ${stepUuid} found in`, byteStepsMap);
-  }
-  const stepSubmission = getStepSubmission(stepUuid);
-
-  const allQuestionsAnswered = step.stepItems
-    .filter(isQuestion)
-    .every((question) => (stepSubmission?.itemResponsesMap?.[question.uuid] as string[] | undefined)?.length);
-
-  return allQuestionsAnswered;
-}
-
-export function updateItemValue(
-  stepUuid: string,
-  itemUuid: string,
-  itemValue: string | string[],
-  setByteSubmission: (fn: (submission: TempByteSubmission) => TempByteSubmission) => void
-) {
-  setByteSubmission((prevByteSubmission: TempByteSubmission) => {
-    return {
-      ...prevByteSubmission,
-      stepResponsesMap: {
-        ...prevByteSubmission.stepResponsesMap,
-        [stepUuid]: {
-          ...prevByteSubmission.stepResponsesMap?.[stepUuid],
-          itemResponsesMap: {
-            ...prevByteSubmission.stepResponsesMap?.[stepUuid]?.itemResponsesMap,
-            [itemUuid]: itemValue,
-          },
-        },
-      },
-    };
-  });
-}
